+++ src/clientity/__init__.py
@@ -2,6 +2,6 @@
 clientity
 """
 __author__ = "Joel Yisrael"
-__version__ = "0.0.0"
+__version__ = "0.1.5"
 
 VERSION = tuple(map(int, __version__.split('.')))