+++ src/clientity/core/grouping/__init__.py
@@ -1,0 +1,4 @@
+# ~/clientity/src/clientity/core/grouping/__init__.py
+from .base import Grouping, WrappedEndpoint, GroupedEndpoint
+from .resource import Resource
+from .namespace import Namespace