+++ src/clientity/core/protocols/__init__.py
@@ -11,3 +11,5 @@ from .interface import (
     SyncInterface,
     AsyncInterface
 )
+
+from .located import Located