+++ src/clientity/core/utils/__init__.py
@@ -1,3 +1,5 @@
 # ~/clientity/src/clientity/core/utils/__init__.py
-from .calls import asynced
-from .http import embody
+from .typers import bound
+from .models import constructors, dictate
+from .calls import synced, asynced, sift
+from .http import embody, domain