+++ src/clientity/core/utils/http.py
@@ -2,12 +2,14 @@
 from __future__ import annotations
 import typing as t
 
+from urllib.parse import urlparse
+
 from clientity.logs import log
-from clientity.core.hints import Requesting, Embodied
+from clientity.core.hints import Requested, Embodied
 from clientity.core.protocols import Requestable
 
 
-def embody(obj: Requesting) -> t.Optional[Embodied]:
+def embody(obj: Requested) -> t.Optional[Embodied]:
     if obj is None: return None
 
     if isinstance(obj, Requestable):
@@ -23,3 +25,11 @@ def embody(obj: Requesting) -> t.Optional[Embodied]:
     # TODO: handle known libraries e.g. dataclasses, pydantic
 
     return "json", obj
+
+
+def domain(url: str) -> str:
+    parsed = urlparse(url)
+    splits = parsed.path.split('/')[0]
+    value = (parsed.netloc or splits)
+    log.debug(f"(domain) parsed {url} -> {value}")
+    return value